ref(normalization): Move span validation to NormalizeProcessor
